requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-core
langgraph
chromadb
sentence-transformers
tiktoken
python-dotenv
structlog
rich
pytest
spacy
nltk
rake-nltk
matplotlib
plotly
langgraph
scikit-learn
rake-nltk
beautifulsoup4
lxml